Reorganizing common utility modules
