+++ src/common/logging.rs
@@ -4,7 +4,7 @@ use log4rs::config::{Appender, Config, Root};
 
 use common::errors::DenebResult;
 
-pub fn init(log_level: LevelFilter) -> DenebResult<()> {
+pub fn init_logger(log_level: LevelFilter) -> DenebResult<()> {
     let stdout = ConsoleAppender::builder().build();
 
     let config = Config::builder().appender(Appender::builder().build("stdout", Box::new(stdout)))
