+++ src/fe/fuse/mod.rs
@@ -1,7 +1,5 @@
 //! Modules which make up the FUSE front-end
 
 mod fs;
-mod params;
 
 pub use self::fs::*;
-pub use self::params::*;
